aster/code/requirements.txt
opencv-python==3.4.2.17
torch==1.1.0
torchvision==0.3.0
scipy==1.1.0
editdistance==0.5.3
matplotlib==3.1.0
pillow==6.1.0
